15/08/18

+++ api/SmartCity3/Models/Announcement.cs
@@ -8,8 +8,8 @@ namespace SmartCity3
         public int Id { get; set; }
         public DateTime Date { get; set; }
         public string Description { get; set; }
-        public int CoordX { get; set; }
-        public int CoordY { get; set; }
+        public Decimal CoordX { get; set; }
+        public Decimal CoordY { get; set; }
         public int IdStatut { get; set; }
         public int IdAnimal { get; set; }
